Reorganize and expose sizing hooks
