+++ packages/components/src/hooks/fluent8/button.ts
@@ -6,8 +6,8 @@ import type { IButtonProps, IButtonStyles, IIconProps } from '@fluentui/react'
 import merge from 'lodash-es/merge.js'
 import { useMemo } from 'react'
 
-import type { Size } from '../types.js'
 import { SMALL_FONT_SIZE } from './constants.js'
+import type { Size } from './types.js'
 
 export function useButtonProps(
 	props: Partial<IButtonProps>,
