+++ packages/components/src/hooks/fluent8/spinbutton.ts
@@ -11,8 +11,8 @@ import type {
 import merge from 'lodash-es/merge.js'
 import { useMemo } from 'react'
 
-import type { Size } from '../types.js'
 import { SMALL_FONT_SIZE, SMALL_INPUT_HEIGHT } from './constants.js'
+import type { Size } from './types.js'
 
 export function useSpinButtonProps(
 	props: Partial<ISpinButtonProps>,
